Add docker-publish workflow && Update README && Some improvements
