+++ .github/workflows/gh-release.yml
@@ -28,8 +28,6 @@ jobs:
           tag_name: ${{ github.ref }}
           release_name: Release ${{ github.ref }}
           body: |
-            Changes in this Release
-            - First Change
-            - Second Change
+            See ChangeLog
           draft: true
           prerelease: false
