+++ docker/Dockerfile
@@ -12,10 +12,9 @@ RUN pip install --no-cache-dir  \
     "python-dotenv" \
     "httpx"
 
-WORKDIR /home
 COPY . /home/openai-forward
 WORKDIR /home/openai-forward
 
-RUN pip install . --no-cache-dir && rm -rf /home/openai-forward/*
+RUN pip install . --no-cache-dir && rm -rf /home/openai-forward
 EXPOSE 8000
 ENTRYPOINT ["openai_forward", "run"]
