+++ openai_forward/__init__.py
@@ -1,1 +1,7 @@
-__version__ = "0.0.9"
+__version__ = "0.1.0"
+
+from dotenv import load_dotenv
+from .config import setting_log
+
+load_dotenv()
+setting_log(log_name="openai_forward.log")
