+++ openai_forward/_base.py
@@ -1,20 +1,13 @@
-import orjson
-from orjson import JSONDecodeError
 from fastapi import Request
 from fastapi.responses import StreamingResponse, RedirectResponse, FileResponse
 from loguru import logger
 import httpx
-from httpx._decoders import LineDecoder
 from starlette.background import BackgroundTask
 import os
-from dotenv import load_dotenv
 from sparrow import yaml_dump, yaml_load
 from pathlib import Path
 from threading import Thread
-from .config import setting_log
-
-load_dotenv()
-setting_log(log_name="openai_forward.log")
+from .content.chat import log_chat_completions
 
 
 class OpenaiBase:
@@ -44,7 +37,7 @@ class OpenaiBase:
         if len(chat_info_list) >= save_threshold:
             cls._chat_file_idx += 1
             chat_file = f"chat_{cls._chat_file_idx}.yaml"
-        if len(chat_info_list) >= save_threshold * save_freq*cls._save_freq_idx:
+        if len(chat_info_list) >= save_threshold * save_freq * cls._save_freq_idx:
             cls._save_freq_idx += 1
             yaml_dump(chat_file, chat_info_list)
             cls.chat_info_list = []
@@ -61,43 +54,9 @@ class OpenaiBase:
         else:
             return False
 
-    @staticmethod
-    def _parse_iter_line_content(line: str):
-        line = line[6:]
-        try:
-            line_dict = orjson.loads(line)
-            return line_dict['choices'][0]['delta']['content']
-        except JSONDecodeError:
-            return ""
-        except KeyError:
-            return ""
-
     @classmethod
     def log_chat_completions(cls, bytes_: bytes):
-        decoder = LineDecoder()
-        txt_lines = decoder.decode(bytes_.decode('utf-8'))
-        line0 = txt_lines[0]
-        target_info = dict()
-        if line0.startswith("data:"):
-            line0 = orjson.loads(line0[6:])
-            msg = line0['choices'][0]['delta']
-        else:
-            line0 = orjson.loads(line0)
-            msg = line0['choices'][0]['message']
-
-        target_info['created'] = line0['created']
-        target_info['id'] = line0['id']
-        target_info['model'] = line0['model']
-        target_info['role'] = msg['role']
-        target_info['content'] = msg.get("content", "")
-        # loop for stream
-        for line in txt_lines[1:]:
-            if line in ("", "\n", "\n\n"):
-                continue
-            elif line.startswith("data: "):
-                target_info['content'] += cls._parse_iter_line_content(line)
-            else:
-                logger.warning(f"line not startswith data: {line}")
+        target_info = log_chat_completions(bytes_)
         logger.info(f"{target_info}")
         cls._current_chat_info.append({target_info['role']: target_info['content']})
         Thread(target=cls.save_chat_info).start()
